Seperate general module into mute and session_id_request
